Convert `enable_if` guard to `static_assert`
The `std::enable_if` guard is more suited to overload selection when
there might be some ambiguity in the overload selection. The
`static_assert` is better when the method being called is fixed, and it
simply needs to assert it is being used correctly with a compatible
type. The `static_assert` will provide a better error message when used
with an incompatible type.