Validate PE signature offset
For a valid PE file, the PE signature offset must point within the file.

Perhaps we should validate the DOS header before reading the PE offset.
That would help ensure we are reading a valid PE offset rather than
random gargabe, which may or may not point within the file. Though it's
not clear from the documentation if the full DOS header is required, or
just the PE offset.

This checks results in a better error message when processing some non
DLL files.